+++ lib/glimmer-dsl-opal/samples/hello/hello_checkbox_group.rb
@@ -19,6 +19,9 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+# This sample demonstrates the use of a `checkbox_group` (aka `check_group`) in Glimmer, which provides terser syntax
+# than HelloCheckbox for representing multiple checkbox buttons (`button(:check)`) by relying on data-binding to
+# automatically spawn the `checkbox` widgets (`button(:check)`) based on available options on the model.
 class HelloCheckboxGroup
   class Person
     attr_accessor :activities
@@ -36,11 +39,13 @@ class HelloCheckboxGroup
     end
   end
   
-  include Glimmer
+  include Glimmer::UI::CustomShell
   
-  def launch
-    person = Person.new
-    
+  before_body {
+    @person = Person.new
+  }
+  
+  body {
     shell {
       text 'Hello, Checkbox Group!'
       row_layout :vertical
@@ -51,18 +56,18 @@ class HelloCheckboxGroup
       }
       
       checkbox_group {
-        selection <=> [person, :activities]
+        selection <=> [@person, :activities]
       }
     
       button {
         text 'Reset Activities'
         
         on_widget_selected do
-          person.reset_activities
+          @person.reset_activities
         end
       }
-    }.open
-  end
+    }
+  }
 end
 
-HelloCheckboxGroup.new.launch
+HelloCheckboxGroup.launch